code moved from main to server
