+++ main.go
@@ -2,9 +2,7 @@ package main
 
 import (
 	"flag"
-	"fmt"
 	"log"
-	"net/http"
 	"os"
 	"strings"
 
@@ -61,8 +59,6 @@ func init() {
 }
 
 func main() {
-	var err error
-
 	opts := &imagejob.ServerOpts{
 		Port:                viper.GetString("port"),
 		Domain:              viper.GetString("domain"),
@@ -93,31 +89,5 @@ func main() {
 		opts.StorageDriver = storage.NewFileDriver(opts.FSBase)
 	}
 
-	// semaphores control concurrent http client requests
-	imagejob.SpecificThrotling = make(map[string]chan struct{}, 20)
-	imagejob.GlobalThrotling = make(chan struct{}, opts.MaxRequest)
-
-	mux := &imagejob.Mux{
-		Routes: make(map[string]func(http.ResponseWriter, *http.Request)),
-	}
-	mux.Handle("/robots.txt", imagejob.Middleware(imagejob.RobotsTXT, opts))
-	mux.Handle("/up", imagejob.Middleware(imagejob.Up, opts))
-	mux.Handle("/image/fetch/", imagejob.Middleware(imagejob.Fetch(opts), opts))
-	server := http.Server{
-		Addr:    ":" + opts.Port,
-		Handler: mux,
-	}
-
-	if SSLDir := viper.GetString("ssl_dir"); SSLDir == "" {
-		fmt.Println("Listening on port", opts.Port)
-		err = server.ListenAndServe()
-	} else {
-		fmt.Println("Listening with SSL on port", opts.Port)
-		err = server.ListenAndServeTLS(SSLDir+"server.pem", SSLDir+"server.key")
-	}
-
-	if err != nil {
-		log.Fatal("ListenAndServe cannot start: ", err)
-		raven.CaptureError(err, nil)
-	}
+	imagejob.Serve(opts)
 }
